feat(expression-compatibility): remove feature flags and use `ffi` instead.
This commit contains two parts:
1. Moving everything guarded by feature flag `expr_validation` to `ffi.rs`.
2. Remove feature flag `expr_validation`.
3. Remove public interfaces `get_field` and `get_operator` for `Predicate`.

No internal logic has been changed.
